@@ -23,7 +23,7 @@ classifiers = [
 keywords = ["airbyte", "connector-development-kit", "cdk"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 airbyte-protocol-models-pdv2 = "^0.12.2"
 backoff = "*"
 cachetools = "*"
