@@ -7,14 +7,17 @@ from __future__ import (  # Used to evaluate type hints at runtime, a NameError:
 )
 
 import time
+from copy import copy
 from typing import Any, List, MutableMapping
 
 from airbyte_cdk.models import AirbyteControlConnectorConfigMessage, AirbyteControlMessage, AirbyteMessage, OrchestratorType, Type
 
 
-class ObservedDict(dict):
-    def __init__(self, non_observed_mapping: MutableMapping, observer: ConfigObserver, update_on_unchanged_value=True) -> None:
-        non_observed_mapping = non_observed_mapping.copy()
+class ObservedDict(dict):  # type: ignore # disallow_any_generics is set to True, and dict is equivalent to dict[Any]
+    def __init__(
+        self, non_observed_mapping: MutableMapping[Any, Any], observer: ConfigObserver, update_on_unchanged_value: bool = True
+    ) -> None:
+        non_observed_mapping = copy(non_observed_mapping)
         self.observer = observer
         self.update_on_unchanged_value = update_on_unchanged_value
         for item, value in non_observed_mapping.items():
@@ -29,7 +32,7 @@ class ObservedDict(dict):
                         value[i] = ObservedDict(sub_value, observer)
         super().__init__(non_observed_mapping)
 
-    def __setitem__(self, item: Any, value: Any):
+    def __setitem__(self, item: Any, value: Any) -> None:
         """Override dict.__setitem__ by:
         1. Observing the new value if it is a dict
         2. Call observer update if the new value is different from the previous one
@@ -58,7 +61,7 @@ class ConfigObserver:
         emit_configuration_as_airbyte_control_message(self.config)
 
 
-def observe_connector_config(non_observed_connector_config: MutableMapping[str, Any]):
+def observe_connector_config(non_observed_connector_config: MutableMapping[str, Any]) -> ObservedDict:
     if isinstance(non_observed_connector_config, ObservedDict):
         raise ValueError("This connector configuration is already observed")
     connector_config_observer = ConfigObserver()
@@ -67,16 +70,16 @@ def observe_connector_config(non_observed_connector_config: MutableMapping[str, 
     return observed_connector_config
 
 
-def emit_configuration_as_airbyte_control_message(config: MutableMapping):
+def emit_configuration_as_airbyte_control_message(config: MutableMapping[str, Any]) -> None:
     """
     WARNING: deprecated - emit_configuration_as_airbyte_control_message is being deprecated in favor of the MessageRepository mechanism.
     See the airbyte_cdk.sources.message package
     """
     airbyte_message = create_connector_config_control_message(config)
-    print(airbyte_message.json(exclude_unset=True))
+    print(airbyte_message.model_dump_json(exclude_unset=True))
 
 
-def create_connector_config_control_message(config):
+def create_connector_config_control_message(config: MutableMapping[str, Any]) -> AirbyteMessage:
     control_message = AirbyteControlMessage(
         type=OrchestratorType.CONNECTOR_CONFIG,
         emitted_at=time.time() * 1000,
