@@ -85,7 +85,7 @@ class AirbyteTracedException(Exception):
         Prints the exception as an AirbyteTraceMessage.
         Note that this will be called automatically on uncaught exceptions when using the airbyte_cdk entrypoint.
         """
-        message = self.as_airbyte_message().json(exclude_unset=True)
+        message = self.as_airbyte_message().model_dump_json(exclude_unset=True)
         filtered_message = filter_secrets(message)
         print(filtered_message)
 
